[IMP] Move tests to py.test Py.test allow us to run the same tests for a list of configurations. In one run we are therefore able to run the same tests for the browser binding and the atompub binding on alfresco. At this stage, the configurations is only defined for Alfresco, but it's easy to add configurations for others CMIS server.
