+++ src/tests/test_browser.py
@@ -28,7 +28,7 @@ from cmislib.browser.binding import BrowserACE
 from cmislib.browser.binding import BrowserACL
 
 
-class BrowserACLTest(unittest.TestCase):
+class TestBrowserACL(unittest.TestCase):
 
     def setUp(self):
         self.aceUser1 = BrowserACE(
@@ -42,56 +42,54 @@ class BrowserACLTest(unittest.TestCase):
         originalEntries = self.acl.originalEntries
         for entry in [self.aceUser1, self.aceUser2]:
             copy = originalEntries.get(entry.principalId)
-            self.assertTrue(copy)
+            assert copy
             # check we have 2 different instances of the same object
-            self.assertNotEquals(id(entry), id(copy))
-            self.assertEqual(entry, copy)
+            assert id(entry) != id(copy)
+            assert entry == copy
 
     def test_get_removed_aces(self):
         # test the complete removal of an alc entry
         self.acl.removeEntry(self.aceUser1.principalId)
         removedAces = self.acl.getRemovedAces()
-        self.assertEqual(len(removedAces), 1)
-        self.assertEqual(removedAces[0], self.aceUser1)
+        assert len(removedAces) == 1
+        assert removedAces[0] == self.aceUser1
         # test partial removal of an entry (delete all + add an existing one
         #  with same direct)
         self.acl.removeEntry(self.aceUser2.principalId)
         self.acl.addEntry(self.aceUser2.principalId, 'cmis:write',
                           direct=False)
         removedAces = self.acl.getRemovedAces()
-        self.assertEqual(len(removedAces), 2)
+        assert len(removedAces) == 2
         toCheck = None
         for removedAce in removedAces:
             if removedAce.principalId == self.aceUser2.principalId:
                 toCheck = removedAce
-        self.assertTrue(toCheck)
-        self.assertEqual(toCheck.principalId, self.aceUser2.principalId)
-        self.assertListEqual(toCheck.permissions, ['cmis:read'])
+        assert toCheck
+        assert toCheck.principalId == self.aceUser2.principalId
+        assert toCheck.permissions == ['cmis:read']
 
     def test_get_added_aces(self):
         # add new entry for a new princpal
         self.acl.addEntry('user3', 'cmis:all')
         addedAces = self.acl.getAddedAces()
-        self.assertEqual(len(addedAces), 1)
-        self.assertEqual(addedAces[0], BrowserACE('user3', 'cmis:all', True))
+        assert len(addedAces) == 1
+        assert addedAces[0] == BrowserACE('user3', 'cmis:all', True)
         # add a new entry for the same principal
         self.acl.addEntry(
             'user3', ['cmis:all', 'cmis:write'])
         addedAces = self.acl.getAddedAces()
-        self.assertEqual(len(addedAces), 1)
-        self.assertEqual(
-            addedAces[0],
-            BrowserACE('user3', ['cmis:all', 'cmis:write'], True))
+        assert len(addedAces) == 1
+        assert addedAces[0] == \
+            BrowserACE('user3', ['cmis:all', 'cmis:write'], True)
         # add a new entry for an exising principal
         self.acl.addEntry(
             self.aceUser1.principalId, ['cmis:read','cmis:write'])
         addedAces = self.acl.getAddedAces()
-        self.assertEqual(len(addedAces), 2)
+        assert len(addedAces) == 2
         toCheck = None
         for addedAce in addedAces:
             if addedAce.principalId == self.aceUser1.principalId:
                 toCheck = addedAce
-        self.assertTrue(toCheck)
-        self.assertEqual(
-            toCheck,
-            BrowserACE(self.aceUser1.principalId, ['cmis:write'], True))
+        assert toCheck
+        assert toCheck == \
+            BrowserACE(self.aceUser1.principalId, ['cmis:write'], True)
